added a fuzzy_ai_taggable class to be able to tag items and karts with a generic method
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/branches/ai@10419 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ trunk/src/karts/controller/player_controller.hpp
@@ -53,7 +53,7 @@ private:
     //unsigned int   m_times_average_rank_was_computed; // Used only by exact average formula
     float          m_timer;
     
-    unsigned int   m_crash_count; // times the player crashed within the 20 last sec.
+    float          m_crash_count; // times the player crashed within the 20 last sec.
     float          m_old_speed;
     //  -- End of fuzzy AI variables --
